Trial multi-file support
For the case where spark has written data without a _metadata file
