+++ fastparquet/test/test_api.py
@@ -142,3 +142,15 @@ def test_request_nonexistent_column(tempdir):
     pf = ParquetFile(fn)
     with pytest.raises(ValueError):
         pf.to_pandas(columns=['y'])
+
+
+def test_read_multiple_no_metadata(tempdir):
+    df = pd.DataFrame({'x': [1, 5, 2, 5]})
+    write(tempdir, df, file_scheme='hive', row_group_offsets=[0, 2])
+    os.unlink(os.path.join(tempdir, '_metadata'))
+    import glob
+    flist = glob.glob(os.path.join(tempdir, '*'))
+    pf = ParquetFile(flist)
+    assert len(pf.row_groups) == 2
+    out = pf.to_pandas()
+    pd.util.testing.assert_frame_equal(out, df)
